+++ src/pascalc/lexer.c
@@ -449,6 +449,8 @@ void next_token(void) {
         else if (strcasecmp(token.text, "file") == 0) token.type = TOKEN_FILE_TYPE;
         else if (strcasecmp(token.text, "seek") == 0) token.type = TOKEN_SEEK;
         else if (strcasecmp(token.text, "filesize") == 0) token.type = TOKEN_FILESIZE;
+        else if (strcasecmp(token.text, "blockread") == 0) token.type = TOKEN_BLOCKREAD;
+        else if (strcasecmp(token.text, "blockwrite") == 0) token.type = TOKEN_BLOCKWRITE;
         else if (strcasecmp(token.text, "sizeof") == 0) token.type = TOKEN_SIZEOF;
         else if (strcasecmp(token.text, "abstract") == 0) token.type = TOKEN_ABSTRACT;
         else if (strcasecmp(token.text, "sealed") == 0) token.type = TOKEN_SEALED;